requirements.txt
numpy==1.26.4
ninja==1.11.1.1
torch==2.5.1
triton==3.0.0
einops==0.8.0
liger-kernel==0.2.1
tiktoken==0.7.0
transformers==4.44.0
datasets==2.20.0
safetensors==0.4.4
wandb==0.19.0
psutil==6.0.0
matplotlib==3.9.2
ipykernel==6.29.5
colorama==0.4.6
orjson==3.10.12
pynvml==12.0.0
nvidia-ml-py==12.570.86
plotly==6.0.0
causal-conv1d==1.5.0.post8
mamba-ssm==2.2.4
